Creating inscribed triangle
Using a right triangle with the height of 2 and a base of 1 we get the area of 1. With that r,s,t are all vertices of the inscribed triangle that are within the main right triangle. I got it to work with one inscribed triangle. This is all done in serial, no MPI yet